# backend/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: YSpotify
  version: '1.0'
paths:
  /api/groups:
    get:
      tags:
        - Group
      summary: Get all groups
      responses:
        '200':
          description: Returns the public Info of the Group
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicGroupInfo'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicGroupInfo'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicGroupInfo'
        '204':
          description: If there are no Groups
  /api/groups/current:
    get:
      tags:
        - Group
      summary: Get the group the authenticated User belongs to.
      responses:
        '200':
          description: Returns the Group Members List
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberInfo'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberInfo'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberInfo'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: If the User is not in a Group
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/groups/join/{groupName}':
    post:
      tags:
        - Group
      summary: Join a group by the group name
      parameters:
        - name: groupName
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns public Info of the Group the User joined
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            application/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            text/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/groups/leave:
    post:
      tags:
        - Group
      summary: Leave a group
      responses:
        '200':
          description: Returns public Info of the Group the User left
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            application/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            text/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: If the User is not in a Group
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/groups/synchronize:
    post:
      tags:
        - Group
      summary: "As the Group Leader, Synchronize all the Group's Members' Players to the current\r\nPlayback state of the Group Leader."
      responses:
        '200':
          description: Returns public Info of the Group the User left
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            application/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
            text/json:
              schema:
                $ref: '#/components/schemas/PublicGroupInfo'
        '401':
          description: If the User is not Authenticated
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: If the User is not the Group Leader
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: If A resource necessary to Synchronize could not be found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
...
components:
  schemas:
    GroupMemberInfo:
      type: object
      properties:
        username:
          type: string
          nullable: true
        is_group_leader:
          type: boolean
        spotify_user:
          $ref: '#/components/schemas/SpotifyUserInfo'
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    PublicGroupInfo:
      type: object
      properties:
        name:
          type: string
          nullable: true
        user_count:
          type: integer
          format: int32
      additionalProperties: false
    SpotifyUserInfo:
      type: object
      properties:
        username:
          type: string
          nullable: true
        currentlyListeningTo:
          $ref: '#/components/schemas/TrackSummary'
        deviceName:
          type: string
          nullable: true
      additionalProperties: false
    TrackSummary:
      type: object
      properties:
        name:
          type: string
          description: The Name of the Song
          nullable: true
        artistName:
          type: string
          description: The Name of the Artist
          nullable: true
        albumName:
          type: string
          description: The Name of the Album the Song belongs to
          nullable: true
      additionalProperties: false
      description: Model representing a Spotify Song
...
  securitySchemes:
    Bearer:
      type: http
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]